requirements.txt
docopt==0.6.2
jsonschema==2.5.1
python-frontmatter==0.2.1
PyYAML==3.11
requests==2.9.1
six==1.10.0
-e .